moved calibration curve interpolation
